Testing MACDSuggestionAnalyzer

+++ Market/Split.cs
@@ -4,7 +4,7 @@ namespace Market
 {
     public class Split
     {
-        public int Key { get; set; }
+        public Int64 Key { get; set; }
         public int StockKey { get; set; }
         public DateTime TimeStamp { get; set; } 
         public double SplitRatio { get; set; }
